Node-RED integration: Added Broker Node - broker.html: Contains node's appearance and properties - broker.js: Contains node's functionality - package.json: Added broker to nodes
Note / Bug: Broker doesn't send out message
If a node is instantiated or dragged onto the canvas before the Broker, it won't receive its message.
This is because the message is sent out once using the constructor upon node construction.
Solutions: Insert a delay into broker or Send the message out periodically